Add Weather service and geocoding service

+++ src/main/java/com/example/demo/dto/WeatherDataInfo.java
@@ -1,4 +1,6 @@
 package com.example.demo.dto;
 
-public record WeatherDataInfo(int Temperature, int high, int low, String extendedForecast) {
+public record WeatherDataInfo(double Temperature, double high, double low, String extendedForecast) {
+
+
 }
